+++ script/test/CHRPresale.testnet.deploy.s.sol
@@ -9,10 +9,10 @@ contract CHRPresaleTestnetDeployScript is Script {
     address mockUSDT = 0x6b423B7Dd9c36AeDcF16525e6Eb827c7a3a4FA11;
     address mockAggregator = 0x1E5CAdBDA5494C8cf1f348e0A64AEb8A66604813;
 
-    uint256 saleStartTime = 1680739200;
-    uint256 saleEndTime = 1688162400;
+    uint32 saleStartTime = 1680739200;
+    uint32 saleEndTime = 1688162400;
 
-    uint256[12] limitPerStage = [
+    uint32[12] limitPerStage = [
         100_000_000,
         190_909_091, // +  90_909_091
         274_242_424, // +  83_333_333
@@ -26,7 +26,7 @@ contract CHRPresaleTestnetDeployScript is Script {
         818_771_403, // + 100_000_000
         961_628_546 // + 142_857_143
     ];
-    uint256[12] pricePerStage = [
+    uint16[12] pricePerStage = [
         10_000,
         11_000,
         12_000,
