+++ test/foundry/contracts/CHRPresale/CHRPresale.helper.t.sol
@@ -14,8 +14,8 @@ contract CHRPresaleHarness is CHRPresale {
         address _usdt,
         uint256 _saleStartTime,
         uint256 _saleEndTime,
-        uint256[12] memory _limitPerStage,
-        uint256[12] memory _pricePerStage
+        uint32[12] memory _limitPerStage,
+        uint16[12] memory _pricePerStage
     ) CHRPresale(_saleToken, _oracle, _usdt, _saleStartTime, _saleEndTime, _limitPerStage, _pricePerStage) {}
 
     /// @notice exposing internal function for testing
@@ -61,7 +61,7 @@ contract CHRPresaleHelper is Test {
 
     uint256 totalSupply = 1_000_000_000;
 
-    uint256[12] limitPerStage = [
+    uint32[12] limitPerStage = [
         100_000_000,
         190_909_091, // +  90_909_091
         274_242_424, // +  83_333_333
@@ -75,7 +75,7 @@ contract CHRPresaleHelper is Test {
         818_771_403, // + 100_000_000
         961_628_546 // + 142_857_143
     ];
-    uint256[12] pricePerStage = [
+    uint16[12] pricePerStage = [
         10_000,
         11_000,
         12_000,
